# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Deskripsi API",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/auth/authenticated": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Endpoint untuk halaman authenticated",
        "description": "endpoint untuk halaman authenticated",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}